+++ src/components/SignUp/SignUp.tsx
@@ -5,13 +5,8 @@ import useUserContext from "../../hooks/useUserContext";
 import Sign from "../Sign/Sign";
 import SignTitle from "../SignTitle/SignTitle";
 import SignForm from "../SignForm/SignForm";
-import { IObjectValues } from "../../hooks/useFormValid";
-
-type SignUpProps = {
-  onRegistration: (inputValues: IObjectValues) => void;
-  message: string;
-  isLoading: boolean;
-}
+import { InputValuesType } from "../Types/InputValuesType";
+import { SignUpProps } from "../Types/props.types";
 
 function SignUp({ onRegistration , message, isLoading}: SignUpProps) {
   const { loggetIn } = useUserContext();
@@ -26,7 +21,7 @@ function SignUp({ onRegistration , message, isLoading}: SignUpProps) {
     }
   },[loggetIn, navigate, pathname])
 
-  const handleSubmit = (inputValues: IObjectValues) => {
+  const handleSubmit = (inputValues: InputValuesType) => {
     onRegistration(inputValues)
   };
   return (
