+++ src/components/Types/ButtonTypes.ts
@@ -1,0 +1,5 @@
+export type ButtonProps = {
+  handleClick: () => void;
+  isOwner?: boolean;
+  place?: string;
+}
