+++ src/components/Types/InputValuesType.ts
@@ -1,0 +1,7 @@
+export type InputValuesType = {
+  name?: string;
+  email?: string;
+  password?: string;
+  shortmovies?: boolean
+  keyword?: string
+}
