+++ src/components/Types/NewMovieTyps.ts
@@ -1,0 +1,13 @@
+export type NewMovieTypes = {
+  nameRU: string,
+  nameEN: string,
+  description: string,
+  country: string,
+  director: string,
+  duration: number,
+  year: string,
+  trailerLink: string,
+  image: string,
+  thumbnail:  string,
+  movieId?: number,
+}
