+++ src/components/Types/currentUser.types.ts
@@ -1,0 +1,6 @@
+export type CurrentUser = {
+  email: string;
+  password: string;
+  name?: string;
+  _id?: string;
+}
