--- src/components/pages/Main/Main.tsx
@@ -1,25 +1,0 @@
-import AboutMe from "../../AboutMe/AboutMe";
-import AboutProject from "../../AboutProject/AboutProject";
-import Footer from "../../Footer/Footer";
-import Header from "../../Header/Header";
-import Portfolio from "../../Portfolio/Portfolio";
-import Promo from "../../Promo/Promo";
-import Techs from "../../Techs/Techs";
-
-function Main() {
-  return ( 
-    <>
-      <Header />
-      <main>
-        <Promo />
-        <AboutProject />
-        <Techs />
-        <AboutMe />
-        <Portfolio />
-      </main>
-      <Footer/>
-    </> 
-  );
-}
-
-export default Main;
