+++ src/contexts/MoviesContext.ts
@@ -5,7 +5,6 @@ interface IMoviesContext {
   addUserMovie: (movie: MovieType) => void;
   savedMoviesList: MovieType[];
   deleteUserMovie:(movieId: string) => void;
-
 }
 
 export const MoviesContext = createContext({} as IMoviesContext);
