+++ src/contexts/currentUserContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react";
-import { CurrentUser } from "../components/App/App";
+import { CurrentUser } from "../components/Types/currentUser.types";
 
 interface ICurrentUser {
   loggetIn: boolean;
