+++ src/utils/config.ts
@@ -39,20 +39,6 @@ export const loadingMessage: {
   profile: 'Сохранение..'
 };
 
-// type handleMovieDataFormatProps = {
-//   nameRU: string
-//   nameEN: string
-//   description: string
-//   country: string
-//   director: string
-//   duration: number
-//   year: number
-//   trailerLink: string
-//   image: string
-//   thumbnail: string
-//   movieId: number
-// }
-
 export function handleMovieDataFormat({ nameRU, nameEN, country, director, duration, year, image, id, description, trailerLink }: MovieType) {
   return {
       nameRU,
